+++ ENSEK.Imports/Dtos/MeterReadingDto.cs
@@ -5,5 +5,5 @@ public class MeterReadingDto
 {
     public int AccountId { get; internal set; }
     public DateTime DateTime { get; internal set; }
-    public double Value { get; internal set; }
+    public decimal Value { get; internal set; }
 }
